chore: split `optimize_children`
