+++ crates/fervid_transform/src/template/core/mod.rs
@@ -1,3 +1,5 @@
+pub mod transform_if;
+pub mod transform_whitespace;
 pub mod v_bind;
 pub mod v_model;
 pub mod v_on;
